+++ src/main/java/com/itesm/panoptimize/repository/AgentsHaveUsersRepository.java
@@ -1,2 +1,8 @@
-package com.itesm.panoptimize.repository;public class AgentsHaveUsersRepository {
+package com.itesm.panoptimize.repository;
+
+import com.itesm.panoptimize.model.AgentsHaveUsers;
+import com.itesm.panoptimize.model.AgentsHaveUsersId;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AgentsHaveUsersRepository extends JpaRepository<AgentsHaveUsers, AgentsHaveUsersId> {
 }
