+++ src/main/java/com/itesm/panoptimize/repository/CompanyRepository.java
@@ -1,2 +1,9 @@
-package com.itesm.panoptimize.repository;public interface CompanyRepository {
+package com.itesm.panoptimize.repository;
+
+import com.itesm.panoptimize.model.Company;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CompanyRepository extends JpaRepository<Company, Long> {
 }
