+++ src/main/java/com/itesm/panoptimize/repository/GlobalMetricRepository.java
@@ -1,2 +1,9 @@
-package com.itesm.panoptimize.repository;public interface GlobalMetricRepository {
+package com.itesm.panoptimize.repository;
+
+import com.itesm.panoptimize.model.GlobalMetric;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface GlobalMetricRepository extends JpaRepository<GlobalMetric, Integer> {
 }
