+++ src/main/java/com/itesm/panoptimize/repository/InstanceRepository.java
@@ -1,2 +1,9 @@
-package com.itesm.panoptimize.repository;public class InstanceRepository {
+package com.itesm.panoptimize.repository;
+
+import com.itesm.panoptimize.model.Instance;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface InstanceRepository extends JpaRepository<Instance, Long>{
 }
