+++ src/main/java/com/itesm/panoptimize/repository/SatisfactionLevelRepository.java
@@ -1,2 +1,9 @@
-package com.itesm.panoptimize.repository;public interface SatisfactionLevelRepository {
+package com.itesm.panoptimize.repository;
+
+import com.itesm.panoptimize.model.SatisfactionLevel;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SatisfactionLevelRepository extends JpaRepository<SatisfactionLevel, Integer> {
 }
