+++ src/main/java/com/itesm/panoptimize/repository/UserRepository.java
@@ -1,2 +1,9 @@
-package com.itesm.panoptimize.repository;public class UserRepository {
+package com.itesm.panoptimize.repository;
+
+import com.itesm.panoptimize.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
 }
