+++ src/main/java/com/itesm/panoptimize/repository/UserTypeRepository.java
@@ -1,2 +1,9 @@
-package com.itesm.panoptimize.repository;public class UserTypeRepository {
+package com.itesm.panoptimize.repository;
+
+import com.itesm.panoptimize.model.UserType;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserTypeRepository extends JpaRepository<UserType, Integer> {
 }
